pyproject.toml
[tool.poetry]
name = "mig3-client"
version = "0.1.0"
description = "Send test result to Mig3 service"
authors = ["Matthew de Verteuil <onceuponajooks@gmail.com>"]
license = "MIT"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Operating System :: Unix",
    "Topic :: Software Development",
    "Topic :: System",
    "Topic :: Terminals",
]

[tool.poetry.scripts]
mig3 = "mig3_client:mig3"

[tool.poetry.dependencies]
click = "^7.0"
pytest = "< 4"
pytest-json = "^0.4.0"
requests = "^2.21"
pathlib2 = "^2.3"
gitpython = "^2.1"
tomlkit = "^0.5.3"

[tool.poetry.dev-dependencies]
pre-commit = "^1.15"
pytest-cov = "^2.6.1"

[tool.black]
line-length = 120
target-version = ['py27', 'py36', 'py37']
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | setup.py
)/
'''

[tool.isort]
force_grid_wrap = 0
include_trailing_comma = true
line_length = 120
multi_line_output = 3
use_parentheses = true
verbose = true

[tool.tox]
legacy_tox_ini = """
[tox]
envlist = py{27,36,37}
skipsdist = true

[testenv]
whitelist_externals =
    poetry
    rm
commands =
    rm -rf pip-wheel-metadata .report.json
    poetry install -v --develop .
    poetry run pytest --cov tests/

[testenv:py27]
basepython = python2
"""

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
